bench: add simple settings benchmarks
